More progress on step 3
It now works for multiple workers, all we need to do now is make sure we can send commands to the workers and also that they're initialised and run in a non-sequential way, where we keep waiting for the response of one worker before we initialise another. We also have to implement the logic for when it's only using one worker, as in that case the worker can just do what the engine did before.